fix: corrigindo erro nas migrations ao utilizar mysql como database

+++ database/migrations/2024_09_15_134551_create_campaigns_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('campaigns', function (Blueprint $table) {
             $table->string('id')->primary();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('address_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('address_id');
             $table->string('Title', 50);
             $table->text('Description');
             $table->json('meta')->nullable();
